+++ core-ng/src/main/java/core/framework/internal/json/JSONAnnotationIntrospector.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.PropertyName;
 import com.fasterxml.jackson.databind.introspect.Annotated;
 import core.framework.api.json.Property;
 
+import java.io.Serial;
 import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.Map;
@@ -14,6 +15,7 @@ import java.util.Map;
  * @author neo
  */
 public class JSONAnnotationIntrospector extends AnnotationIntrospector {
+    @Serial
     private static final long serialVersionUID = 3638140740110527623L;
 
     @Override